+++ infrastructure/lib/authentication-stack.ts
@@ -1,0 +1,37 @@
+import { CfnOutput, Stack, type StackProps } from 'aws-cdk-lib';
+import { UserPool, UserPoolClient, AccountRecovery, Mfa } from 'aws-cdk-lib/aws-cognito';
+import { Construct } from 'constructs';
+
+export class AuthenticationStack extends Stack {
+  public readonly userPool: UserPool;
+  public readonly userPoolClient: UserPoolClient;
+
+  public constructor(scope: Construct, id: string, props?: StackProps) {
+    super(scope, id, props);
+
+    this.userPool = new UserPool(this, 'UserPool', {
+      selfSignUpEnabled: true,
+      signInAliases: {
+        email: true
+      },
+      accountRecovery: AccountRecovery.EMAIL_ONLY,
+      mfa: Mfa.OFF
+    });
+
+    this.userPoolClient = new UserPoolClient(this, 'UserPoolClient', {
+      userPool: this.userPool,
+      generateSecret: false,
+      authFlows: {
+        userPassword: true
+      }
+    });
+
+    new CfnOutput(this, 'UserPoolId', {
+      value: this.userPool.userPoolId
+    });
+
+    new CfnOutput(this, 'UserPoolClientId', {
+      value: this.userPoolClient.userPoolClientId
+    });
+  }
+}